pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "jupancon"
authors = [{name = "Juan Gonz\u00e1lez-Vallinas", email = "juan@multilayer.io"}]
readme = "README.md"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
# I'm aware that this versioning is considered overly strict, 
# but I believe in overly strict package management for a library 
# of this nature.
dependencies = [
    "pandas==1.3.3",
    "sshtunnel==0.4.0",
    "pyparsing==2.4.7", 
    "sqlalchemy-redshift==0.8.6",
    "sqlalchemy-bigquery==1.2.1",
    "google-cloud-bigquery-storage==2.9.1",
    "pyarrow==6.0.0",
    "redshift-connector==2.0.905"
]

[project.optional-dependencies]
test = [
    "pytest==6.2.5",
    "jupyter==1.0.0",
    "jupyterlab==3.1.14",
    "black==21.9b0",
    "flake8==3.9.2",
    "isort==5.9.3",
    "pytest==6.2.5"
]

[project.urls]
Home = "https://github.com/multilayer-io/jupancon"
